+++ gr-fec/lib/ldpc_encoder.cc
@@ -45,7 +45,7 @@ void ldpc_encoder::generic_work(void *inBuffer, void *outBuffer) {
     std::vector<char> inbuf(inputSize);
     memcpy(&inbuf[0], in, inputSize);
     std::vector<char> coded(d_code.encode(inbuf));
-    for(int i=0; i<coded.size();i++){ out[i] = coded[i]; }
+    for(size_t i=0; i<coded.size();i++){ out[i] = coded[i]; }
 }
